Conflicts:
	code/DriveIT/DriveIT.WindowsClient/Image/icons/cars/car.svg
	code/DriveIT/DriveIT.WindowsClient/Image/icons/cars/carquery1.svg
	code/DriveIT/DriveIT.WindowsClient/Image/icons/cars/carquery2.svg
	code/DriveIT/DriveIT.WindowsClient/Image/icons/shared/add.svg
	code/DriveIT/DriveIT.WindowsClient/Image/icons/shared/cancel.svg
	code/DriveIT/DriveIT.WindowsClient/Image/icons/shared/delete.svg
	code/DriveIT/DriveIT.WindowsClient/Image/icons/shared/info.svg
	code/DriveIT/DriveIT.WindowsClient/Image/icons/shared/refresh.svg
	code/DriveIT/DriveIT.WindowsClient/Image/icons/shared/settings.svg
	code/DriveIT/DriveIT.WindowsClient/Image/icons/users/user.svg

+++ code/DriveIT/DriveIT.WebAPI/Controllers/CommentsController.cs
@@ -1,9 +1,11 @@
 ﻿using System.Linq;
 using System.Threading.Tasks;
 using System.Web.Http;
+using DriveIT.Entities;
 using DriveIT.EntityFramework;
 using DriveIT.Models;
 using DriveIT.WebAPI.Models;
+using Microsoft.AspNet.Identity;
 
 namespace DriveIT.WebAPI.Controllers
 {
@@ -20,7 +22,8 @@ namespace DriveIT.WebAPI.Controllers
 
         // GET: api/Comments/5
         // Where 5 is CarId
-        public async Task<IHttpActionResult> Get(int id)
+
+        public async Task<IHttpActionResult> GetByCarId(int id)
         {
             // Remember: It is a car-id. Not commentId.
             var carId = id;
@@ -34,6 +37,7 @@ namespace DriveIT.WebAPI.Controllers
         }
 
         // POST: api/Comments
+        [AuthorizeRoles(Role.Customer)]
         public async Task<IHttpActionResult> Post([FromBody]CommentDto value)
         {
             if (!ModelState.IsValid)
@@ -44,11 +48,17 @@ namespace DriveIT.WebAPI.Controllers
             {
                 return BadRequest("Null value not allowed.");
             }
+            if (User.Identity.GetUserId() != value.CustomerId)
+            {
+                return BadRequest("CustomerId should be the same as the logged in user");
+            }
+
             var newCommentId = await _repo.CreateComment(value.ToEntity());
             return CreatedAtRoute("DefaultApi", new { id = newCommentId }, value);
         }
 
         // PUT: api/Comments/5
+        [AuthorizeRoles(Role.Customer)]
         public async Task<IHttpActionResult> Put(int id, [FromBody]CommentDto value)
         {
             if (!ModelState.IsValid)
@@ -60,17 +70,30 @@ namespace DriveIT.WebAPI.Controllers
             {
                 return NotFound();
             }
+            if (User.Identity.GetUserId() != comment.CustomerId)
+            {
+                return BadRequest("Customer cannot change other customers comments.");
+            }
+            if (User.Identity.GetUserId() != value.CustomerId)
+            {
+                return BadRequest("Customer cannot set author to another person.");
+            }
             await _repo.UpdateComment(id, value.ToEntity());
             return Ok();
         }
 
         // DELETE: api/Comments/5
+        [AuthorizeRoles(Role.Customer, Role.Administrator)]
         public async Task<IHttpActionResult> Delete(int id)
         {
             var comment = await _repo.GetCommentWithId(id);
             if (comment == null)
             {
                 return NotFound();
+            }
+            if (User.IsInRole(Role.Customer.ToString()) && User.Identity.GetUserId() != comment.CustomerId)
+            {
+                return BadRequest("Customer cannot delete others comments.");
             }
             await _repo.DeleteComment(id);
             return Ok();
